Project Service project fix
