+++ .gitignore
@@ -50,3 +50,6 @@ out/
 /dist/
 /nbdist/
 /.nb-gradle/
+
+### Google Calendar ###
+!**/src/main/resources/calendar
