--- src/main/java/faang/school/projectservice/exception/EntityNotFoundException.java
@@ -1,7 +1,0 @@
-package faang.school.projectservice.exception;
-
-public class EntityNotFoundException extends RuntimeException{
-    public EntityNotFoundException(String message) {
-        super(message);
-    }
-}
