+++ src/main/java/faang/school/projectservice/validator/ProjectValidator.java
@@ -1,7 +1,5 @@
 package faang.school.projectservice.validator;
 
-import faang.school.projectservice.exception.DataValidationException;
-import faang.school.projectservice.service.ProjectService;
 import faang.school.projectservice.dto.project.ProjectDto;
 import faang.school.projectservice.exception.DataValidationException;
 import faang.school.projectservice.model.Project;
@@ -17,7 +15,6 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class ProjectValidator {
 
-    private final ProjectService projectService;
     private final ProjectRepository projectRepository;
     private final String VALIDATION_MASSAGE_NAME_OF_PROJECT_IS_EXISTS =
             "The owner with id:%d already has a project with  name:%s";
@@ -43,18 +40,6 @@ public class ProjectValidator {
             String errMessage = String.format("Project with ID: %d already has calendar", project.getId());
             log.error(errMessage);
             throw new IllegalStateException(errMessage);
-        }
-    }
-
-    public void checkUserPermission(long projectId, long userId) {
-
-        boolean campaignCreateByProjectOwner = projectService.checkOwnerPermission(userId, projectId);
-        boolean campaignCreateByManager = projectService.checkManagerPermission(userId, projectId);
-
-        if (!campaignCreateByProjectOwner && !campaignCreateByManager) {
-            log.warn("You don't have permission to manage campaign for project with id: {}. " +
-                    "Only project owner or manager can manage campaign", projectId);
-            throw new DataValidationException("You don't have permission to manage the campaign");
         }
     }
 }
